// part9/flight-diary/src/types.ts
import { z } from 'zod';
import { NewDiaryEntrySchema } from './services/diaryService';

export enum Weather {
  Sunny = 'sunny',
  Rainy = 'rainy',
  Cloudy = 'cloudy',
  Stormy = 'stormy',
  Windy = 'windy',
}

export enum Visibility {
  Great = 'great',
  Good = 'good',
  Ok = 'ok',
  Poor = 'poor',
}

export interface DiaryEntry {
  id: number;
  date: string;
  weather: Weather;
  visibility: Visibility;
  comment?: string;
}

export type NewDiaryEntry = z.infer<typeof NewDiaryEntrySchema>;

export type NonSensitiveDiaryEntry = Omit<DiaryEntry, 'comment'>;